requirements.txt
groq
gradio
python-dotenv
gTTS
pyglet
